requirements.txt
Django==1.11.21
django-filter==2.0.0
djangorestframework==3.9.4
Markdown==3.1.1
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
